requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
sentence-transformers